Moved D-Bus signal handlers to own file and lib.
This is for a less crammed dbus_iface.c and to enable testing.

In addition, the new dbus_handlers.cc file implements the step away
from C code with GLib dependencies to pure C++ domain. This is also
the reason why the new header file dbus_handlers.h is not called
dbus_handlers.hh---to make it look fine to include it from a pure C
file.