Minimal implementation of Dcpd Playback signals.
The signals defined by the de.tahifi.Dcpd.Playback interface are now
handled. All of them end up as calls of ViewManagerIface functions.

To accomplish this, the low-level D-Bus C code now takes an opaque
void pointer to a ViewManagerIface instance and passes it on to the
GLib handlers. These are actually C++ functions that can use the
ViewManagerIface.