List navigation filtering (selectability, visibility).
Using the list item flags, visibility and selectability can by
dynamically controlled. To this end, the user of a List::Nav object
must pass a List::NavItemFilterIface object when constructing it. The
latter implements a few functions that determine visibility and
selectability for each list item at runtime, and the List::Nav object
is using these functions to skip invisible items and avoid selection
of unselectable items.

The unit tests show that at least basic selectability filtering is
currently working, but some more tests are missing.