Automatically test with Valgrind.
If Valgrind is detected, then ``make check'' will run tests twice:
once as usual, and once with Valgrind.

The usual run is used to obtain the Cutter report. These Cutter tests
are now executed without the --keep-opening-modules option.

The Valgrind run keeps the --keep-opening-modules option in place to
enable memory leak detection. It does not generate a Cutter report,
but instead a Valgrind XML file is produced.