Removed redundant declarations from all unit tests.
There were forward declarations of all tests in namespaces at the
beginning of the file. This was supposed to provide a quick overview
of tests and to keep the comments out of the implementation. What
seemed to be at first a good idea in theory turned out to be a
constant source of mistake and a growing maintenance burden.

It is not possible to use the -Wmissing-prototypes warning with C++,
so spelling a function's name differently in the block of static
forward declaration that carry the documentation from the name used
way further down where the function is defined yields no compiler
warning. Consequently, it was also possible to write a test, but to
forget the forward declaration with its comment, unnoticed. Such
mistakes happened all the time, and constantly checking for these
issues by hand is just stupid.

Further, the documentation for tests and their implementation turned
out to be too far apart. Jumping back and forth between them was
necessary to check if they are in sync, and this was annoying.