Require glib and gio-unix >= 2.36 for g_unix_fd_add().
It's not clear which "package" is supposed to contain the
g_unix_fd_add() function (in the end it's simply going to be Debian
package "libglib2.0" anyway, containing all that's needed), so the
version check is tacked to both now.

And no, I'm not going to reimplement that missing function only
because Debian/Raspbian folks think sticking with ancient software
versions is good for users (glib 2.33 in this case).