Finally sorted out language switching at runtime.
GNU gettext() and glibc and quite confusing in their way of using the
various locale-related environment variables. Turns out that LANGUAGE
has highest precedence, but documentation for this is extremely hard
to find.

Our code now only considers LC_ALL, neither LANGUAGE nor LANG. If
LC_ALL is not set, then it is set to some default value.

Tested on Ubuntu only so far, may or may not work on Raspbian.