Generic generation of DRCP XML in ViewIface.
Member functions serialize() and update() are not pure anymore. Their
implementations call a simple function that starts a DCP transaction,
calls three virtual functions for writing the XML output, and commits
or aborts the transaction. Distinction between full serialization and
update is done by a flag.

There are basic implementations for the XML-writing functions, from
which the one responsible for writing most part of the XML body is
likely going to be overridden by most classes derived from ViewIface.