Unit tests for distance-from-end-of-list functions.
It is possible to query the distance of the selection from the top and
from the bottom item of the displayed list. This is used by the page
up and page down navigation to make the first jump behave the way a
user would probably expect it to (i.e., when in the middle of a list,
page up/down does not scroll, but jumps to the first/last visible
item).