Fix failing unit tests.
It is now possible to determine the total number of visible items by
asking the underlying filter. This information is used to correctly
compute the distance to top/bottom of the displayed list.

In addition, the function for getting the total number of visible
items has been added to the public interface of NavItemFilterIface, so
yet another unit test was added as well.