Interfaces for getting line number of selected item.
For DRCP browse views, the list navigation object needs to be able to
return the currently selected line number. To this end, we add
List::Nav::get_line_number_by_cursor() and the more generic
List::Nav::get_line_number_by_item(). The latter of these functions
may turn out useless in the future, in which case the former, more
specialized function could probably be implemented a bit more
efficiently for filtered lists---that's why there are two interfaces.

The implementation is currently broken, but there are a bunch of unit
tests that should be correct.