Grep for extra errors in unit test log file.
Some errors are not caught by the Automake test driver because Cutter
won't throw a proper error. Specifically, when trying to use an
undefined symbol as a result of forgetting to linked against some
object or forgetting to provide an implementation, Cutter will only
show a warning (!), but otherwise carry on with the next test and will
not error out. What a horrible, horrible idea...