Fix DCPD connection timeout code.
GLib function g_source_remove() actually destroys the event source, so
the code attempted to destroy non-existent objects. The behavior of
g_source_remove() is not clearly documented, hence the mistake.

We now create a new timeout event source each time we need it and
remove (destroy) it explicitly when the communication was fast enough
so that the timeout didn't expire. If the timeout expires, the code
relies on GLib to removed it.

See also bc9176da3a487da675016de87ec46977507cff37 (DCPD connection
timeout handling).