Handle failure to start transaction with dcpd.
In case starting serialization fails, a corresponding view signal is
sent. This signal remains active until serialization succeeded or
until the view is hidden, in which case the failure becomes
irrelevant.

We call sched_yield() in case of pending dcpd transactions to give
dcpd a chance to do something. Otherwise we would end up running a
busy loop until the kernel scheduler decides to remove us from the
CPU.

Change-Id: I23bd19a306ef47a768160f999b586cdb184c7812