Remove unneeded ListIface interfaces.
The removed interfaces essentially represented duplicated
functionality already implemented in List::Nav, in particular going up
and down through hierarchies. For the D-Bus-based lists, the required
information is fetched over D-Bus anyway, so the extra in-RAM
hierarchy is redundant at best. Since it wasn't used up to now,
however, it was simply removed.

In case some future list structure needs such information in RAM, it
can be added to the specialized list, with all the tests added again
that got killed by this commit.

Change-Id: I41a611b6fad802feb1a01004c4345ed4f85406b7