Use new List::DBusList for file browsing view.
This new list implementation holds only a small window of the whole
list in RAM. The window is filled on demand over D-Bus.

Internally, List::DBusList uses tested and proven List::RamList to
store the window. Before this commit, the List::RamList class was used
to store the whole list, but fetching big UPnP containers (which is
required to fill a List::RamList) can take several minutes, making the
whole system unusable.

Change-Id: Iead3e4001b7d9de3e32303831fd3fc78a8ab08d8