Partial implementation of playing directories, no shuffle.
All file views are now configured for linear playing now. Streams are
sent to streamplayer's FIFO until it is full. At this point, the
current implementation stops working

In addition to linear directory playing, there is also a shuffle mode
which could, theoretically, be selected. Currently, there is no code
that implements it, and results are undefined for now. Shuffling on
slow media is not that easy, see comments in ticket #27 for some
thoughts on the issues.

Change-Id: I34f7d6c9dcc2c3b47474077d8c876c3ba24ab81f