Avoid hitting an assertion.
The situation in which the assertion failed (and rightly so) occurred
when a list broker sent the invalid list ID 0 without any error code
as a response to a de.tahifi.Lists.Navigation.GetListId(0, 0) request
("give me ID of root directory").

Instead of failing an assertion, thus aborting the program, we now
issue a BUG() warning. This is going to be useful for tracking down
bugs in list brokers.

Change-Id: I9aee7d0b391bf1de36ef522afff56c7bf88a6ed3