Fix Doxygen extraction for ListError.
The ListError class is defined in a file shipped with the
dbus_interfaces submodule. It is based on a C header file which allows
two macros to be defined for overriding default names of the error
enum and enum constants that are defined for the pure C version. The
C++ version defines these macros to avoid the C-style namespaces, but
Doxygen does not seem to extract them correctly.

This change defines the two macros in the Doxyfile.in file so that
automatic and explicit linking to the error codes in documentation
work now.

Change-Id: I72d9d494be2615b535755066ea7de339e92f506d