Remove errno values in some msg_error() calls.
These values often lead to misguiding error messages because the errno
values were expanded to a string and provided the wrong context.
Instead of error codes, in some instances an error code of 0 is passed
to msg_error(), causing the errno string to be suppressed in the log.
In other instances, the error codes are still passed to the logging
function if they fit.

Change-Id: Ief0a9279a4f9a6a7604bd346a9a9a80629ee8846