Add container that stores extra stream information.
For each stream queued into the Streamplayer, extra data is associated
with an ID and stored in the new StreamInfo container. That ID is
going to be sent to Streamplayer so that we can retrieve the stored
information by ID when required.

At the moment, only a fallback title string as read from the list
delivered by some List Broker is stored. In case there is no meta data
inside the stream, this string is used to show the stream on the
device display. Streamplayer simply doesn't (and shouldn't have to)
know the stream name as it appeared in the list, so we store it inside
drcpd.

Change-Id: Id7e33d283fd400825dbb07d829ab712178e04828