Add Player class, used as singleton.
This commit pulls the shared StreamInfo pointer out of the views and
out of the ViewManager class (the latter class basically owned the
StreamInfo object, but it was the wrong place to begin with) and puts
it into the new player. Functionally, this commit does not change
anything, only the organization of code has changed.

There is still a playback state object embedded in each file browser
view. The views, however, do not use these objects by themselves, but
pass them to the player when they need to play something, much like an
opaque token. (An even better approach would be to use a state factory
provided by the player so that views could produce their state when
needed, but not now...)

Change-Id: I7e0153575d53759b2caa3ba99f6d28844f2ee2f4