Improve handling of stream bit rate (#251).
GStreamer sends up to four different tags for the bit rate of the
currently playing stream: "bitrate", "minimum-bitrate",
"maximum-bitrate", and "nominal-bitrate". This commit adds support for
the plain "bitrate" tag and sends one of the four values to the SPI
slave. Function get_bitrate() implements the selection of these tags
based on availability and simple ranking.

Change-Id: I6afdc6392356312a64998c04cb7f7f76287d3b8c