Pull bounce table up to ViewManager::Manager::input().
This makes the DrcpCommand::PLAYBACK_STOP command a globally
recognized command which is always redirected to the play view,
regardless of the active view (before this change, only the browse
views did this redirection).

It seems to make sense to treat the command like this now, but it is
always possible to add view-local bounce tables again later to gain
finer control over commands based on context.

Change-Id: Ia4fe9b48b03d96480becf0378ac41a4f5cb7052c