Add view for managing search parameters.
This "view" is never visible, it only stores the search parameters, if
any, and can send an XML that asks the SPI slave to fill in a query,
and how to.

This version is only an incomplete implementation of a bigger idea.
The only kinds of search queries this implementation can perform is
searching for single strings.

Change-Id: I1d377edf767b6d6fd0bc645fd2c3e03ce7a2a655