New flags for file browser lists: D, u, q, l, and L.
D is for storage "d"evices.
u is for opaque strings ("u"ndefined).
q is for search forms ("q"ueries).
l is for "l"ocked (was L in previous versions).
L is for play "l"ists (was locked in previous versions).

Note that l and L are extra flags now that are used as attributes for
other flags. For instance, it is now possible to distinguish locked
directories from locked files, in theory at least. In practice, our
internal representation of locked items is slightly wrong, preventing
us to emit such flag combinations right now. Hence, locked items are
always emitted as "ul" (opaque string, locked).

Change-Id: Id77f0e6cc6c8d1d31b3d48559763949c0cedcb16