Speed up play view notifications on D-Bus messages.
The view is looked up by name only once, from then on a pointer to the
view is used. The pointer is needed at least once per second, still
the speed-up might, in fact, be negligible. Anyway, the optimization
was obvious and easy to implement. It also makes the D-Bus handlers a
bit easier to read.

Change-Id: I5c1a2603090162c8e0763ad1b5925acf52837251