Add dedicated stream state for buffering.
Buffering state is not derived from the information whether or not the
player is waiting for the start notification. This information is
quite different and causes the play view to show buffering state much
to long.

Starting with this commit, the buffering state must now be retrieved
by calling Playback::PlayerIface::get_assumed_stream_state() instead
of the is_buffering() function member (which has been removed to keep
the interface lean).

Change-Id: I88a2ef780d05aefbb19d2f3f4f6f5a6d67423409