Fix long blocking "Buffering..." message (#290).
Prefetching URLs from TIDAL can take a very long time, and the
corresponding D-Bus method invocation was done synchronously. This
caused all other D-Bus messages to queue up until prefetching was
done.

This commit moves the queuing machinery to a thread. It also turns the
player into a big mess of code that needs to be redesigned now.

Change-Id: Ib2374e5ba737d8008ced564ae1938a565760dd0a