Make use of search button (no functionality yet).
Hitting the blue search button on the remote control is supposed to
open the search box from anywhere. The software is supposed to search
for the "right thing" depending on context.

This patch implements the command routing so that the search button is
wired to the correct places, but the actual search functionality is
not implemented yet.

Change-Id: I33452e9db89aa3c7785c41a36a47c5304bf6c154