Enable quick search in current context within Airable (#235).
Pushing the search button on the remote control now opens the search
dialog that matches the current context. It works by jumping to the
search form list item that is found when entering an Airable top-level
directory for the first time. There is no search for the
root directory because it is not clear what to do in that location.

The changes to enable this are rather intrusive and very specific to
the Airable file browser. Therefore, a new
ViewFileBrowser::AirableView class has been derived from
ViewFileBrowser::View that adds the details of figuring out the search
form links. The ViewFileBrowser::View tries to jump to a search form;
on unexpected failure it aborts the search, if not available it
resorts to binary search of the current list, otherwise it does the
same as it did before when user explicitly selected the search entry.

Change-Id: I84f39c8483d7f8651e527d08e94a51a5e62d2ba7