Simplify and lock down PlayInfo::Data class.
While investigating tickets #334 and #347, the code was a bit hard to
understand in some ways. Turned out that this was caused by the fact
that there were two different ways to set the assumed stream state (by
member function and by direct public member manipulation).

This commit restricts setting the assumed stream state to a
function-only interface which makes it simpler to reason about the
code.

Change-Id: I1be5545decf960326face3f08436f32a408346c9