Asynchronous retrieval of D-Bus list items (#391).
The previously existing synchronous implementation has been rewritten
to make use of the new asynchronous interfaces. Client code had to be
adapted as well to keep everything in shape.

This commit breaks busy state notifications. It is currently not
possible to serialize the busy state update in every situation without
running into deadlocks.

Change-Id: Idd9149d874dc38b30b00c30accb71c0583a4625b