Re-introduce functions based on truly synchronous D-Bus interfaces.
This is needed for finding the search forms in Airable. Mapping this
code to asynchronous interfaces would work as well, of course, but it
is really not required in this context.

The synchronous versions may also come handy in different contexts and
allows to keep some not-so-important areas of the code synchronous for
the time being.

Change-Id: I6eb272d69f01e2711dedeac5e1862e943bd4f4b9