Avoid completely pointless GLib warnings.
Why are they pointless? The warnings warn about NULL being passed to
g_variant_unref(), which is wrong according to API specs (even though
this spec doesn't align with free() or operator delete, by the way).
So they are actually doing a NULL pointer check to emit the ignorable
warning and return from the function, and thus turn NULL into a safe
value. Good work, guys...

The correct behavior would be to assert() or to call abort() to turn
this into an actual error that cannot be ignored to enforce correct
API use.

Change-Id: Ibe4843c471bdbe908cf0e571d13ddd2504e5338e