Fix more race conditions in async D-Bus code (#391).
Fast scrolling through lists was broken and left the user with empty
lists after a while because the scrolling code assumed the whole
cached window is either valid or invalid. This assumption was correct
in the old synchronous approach, but was broken when list fetching
became asynchronous.

This commit adds code for keeping track of the state of cache window
parts. It also requires callers of List::DBusList::get_item_async() to
call List::DBusList::get_item_async_set_hint() beforehand.

Change-Id: I62d3faf2f374e97e9771c87a07e468a5870e2781