Usage permissions per streaming context.
Permissions are implemented as overridable functions that are called
from various locations in the player. We use functions rather that
static tables to allow for dynamic and very specific permissions.

For instance, skipping forward within Deezer programs is only allowed
a certain amount of times per hour, and the server tells whether or
not it is currently allowed. This is very specific to a portion of
Deezer and probably also dependent on the account, impossible to
capture with a static table of permissions.

Change-Id: Iedbd1d7fae65fb4a75110c0716971a990b7cf8b0