Fix synchronization of canceled async ops.
Problem was that an operation canceled via a GLib cancelable is only
really canceled after the GLib main loop has handled the cancel
request, and we get the callback from GLib only at that point. The
callback actually marks the operation as truly canceled. In-place
synchronization with that callback is not possible because control
must flow back to the GLib main loop, and marking the operation as
canceled too soon would lead to trouble.

This commit tweaks the synchronization so that this special condition
is detected and handled by calling the caller of the async op
directly. This is the primary reason why we switched to a recursive
mutex in an earlier commit.

Change-Id: Ib173da397554072adf141b1f390dabea78f0aab1