Restart async list operations on list invalidation.
Occasionally, entering a list by ID fails because the list gets
invalidated after starting the enter operation and before the
operation finishes. The operation then returns an invalid ID error.

Since we are observing the list invalidation signals sent by list
brokers, we can (and now do!) cancel a pending operation that is known
to fail due to a bad list ID and restart it automatically with the new
ID.

This seems to be not a big deal or not very important, but in practice
this case occurred frequently during startup, especially with the UPnP
list broker which may send many list updates while servers get
discovered. The result was an empty server list after startup in most
cases.

Change-Id: Ib2a08521b49c7e3643222f04701694329833dc96