Do not assume that std::move() clears a std::function.
The directory crawler now uses a function template to actually move
callback functions of type std::function from object member to locals
or function parameters, clearing the original object on the way.
Previous versions relied on std::move() to behave like this, which is
generally wrong (because std::move() is basically just a type cast and
behavior depends on the implementation of the move ctor, not on
std::move() itself).

This mistake may or may not have led to funny problems. Skipping still
doesn't work right, though.

Change-Id: I9e752bd68c9ab1c84a705f6a46e80ae9a21397f2