Do not filter play commands in observer mode (#434).
In case a stream was started by external means (smartphone app), the
player is put into a passive observer mode. In this case, play
commands got filtered out in previous versions (pause and stop
commands were not, so this was an inconsistency). This commit always
sends the play command down to the player to fix pause/resume behavior
when playing via app.

There is still one drawback: The play command sent by the app is
indistinguishable from the play command sent by the remote or the
device button. In browse views, the play command will always start
playing whatever is selected, actively. The only way around this is to
make the app play command distinct from the other play command.

Change-Id: I3653e523c8c379122f5bb5887f050451ad0c60ae