Fix processing of multiple active AsyncGetURIs calls.
Rarely, at least two AsyncGetURIs calls can be active at the same
time, the first of which in canceled state. In case an AsyncGetURIs is
canceled and another call is created, it is possible for the first
AsyncGetURIs call being in pending canceled state while the next call
is active, leading to two active AsyncGetURIs instances. This happens
if the main loop has not yet processed the canceled asynchronous D-Bus
call.

Of course, the DBus::AsyncResultAvailableFunction will be called for
each instance in the end. The problem was that this function as
implemented in Playlist::DirectoryCrawler calls yet another callback,
and this callback was reset to nullptr after the call. Since there was
only a single instance of this callback pointer, the second invocation
of the DBus::AsyncResultAvailableFunction function had no callback it
could call, leading to missing upstream notifications.

Change-Id: If9d55f4763b5b0c7100205c12a0130292a6b818f