Remove wrong assertions.
Streamplayer may report a stop notification with error and non-empty
queue even in cases in which the error code would imply an empty
queue. The reason for this is that streamplayer sends the state of the
queue as it was when the message is sent, not as it was when the error
was discovered (error reporting is explicitly decoupled from error
discovery to avoid deadlocks).

Change-Id: Ia241e7cd975d9e0354e0c21f8891ed6654dd7838