Do not rely on next-stream-information from Next() call.
The big problem here is that the information obtained from that call
may be more recent than other information sent by streamplayer. We get
seemingly contradictory information and get out of sync if using that
information.

The reason is that in many cases streamplayer must handle errors in a
context that is different from that in which the error was detected.
It may send updates resulting from errors late, in particular after
answering an intermittent Next() call. Such an answer, however, will
take the most current state after an error has been detected into
account, which will be more recent than what will be sent as a signal
later.

Thus, to keep information in the correct temporal order, we'll solely
rely on signals for status updates.

Change-Id: Iad21d54a1915cb379cb0fe0a63a4e26a5a0c8e57