Specialization of UI::Parameters which includes EventID.
This class is currently unused and was created while going down a
wrong path during development, but it seems to be useful anyway.

In case we ever need to put a single GVariant into different parameter
sets so that the GVariant type is not enough anymore to determine the
correct type, then these parameters need to be wrapped into
UI::SpecificParametersForID objects.

Change-Id: I3d0978fb87b02f873db67a3fba0d470b587b4585