Unfold async list callbacks in player control.
These callbacks were intended as catch-all solutions that do the right
thing depending on context, where context was stored in an explicit
state variable. The context checks, however, made the callback
function quite complicated and hard to follow.

Therefore, the whole crawler context idea has been dropped and
replaced by a bunch of smaller, simpler functions that do the right
thing depending on context. This time, however, context is represented
implicitly by control flow.

A few bugs could be fixed on the spot as the code is a lot clearer
now.

Change-Id: Ic8cf4815e66308a83965d7588206f1de9c188f86