Fix skipping over text file (#493).
Skipping in forward direction seemed to be OK, but reverse direction
was broken. On late error discovery, the crawler recovered by always
proceeding in forward direction. Now we store the direction from which
we have approached each item to be able to switch the crawler into
that direction in case of failure.

Change-Id: I3515247af728909828bf7cadce2a17244565203e