Factor concept of audio source owned by views into own base class.
This commit cleans up the management of association of views with one
or multiple audio sources. The new base class ViewWithAudioSourceBase
adds a few helpers for creating audio sources and registering them
with the audio path manager.

In particular, the ideas mainly added in the following commits are
reworked by this commit:
* dea11ea Decouple view ID from audio source ID.
* 04d31f8 Multiple audio sources per view, primarily for Airable.
* f3af7a2 Explicit representation of an audio source.

Change-Id: Ie0152fdc7c5fb30dc2bd429f0d89c1ad61072875