Make XML serialization a bit more dynamic.
It is possible now to change the default view ID string and screen ID
while the DRCP XML is generated.

To this end, before the XML is generated, the new virtual function
member ViewSerializeBase::about_to_write_xml() is called. This
function is supposed to check what kind of XML output is to be
generated and to return the results of this check as flags in a
uint32_t. These flags are then passed to the various functions used
for XML generation which alter their behavior based on these flags.
This way, the possibly expensive checks need to be performed only
once.

This commit only adds the framework for all of this. There is no code
which makes use of it yet.

Change-Id: I3202fc8c93ba4edd412833614b3a496d6e692e5d