Starting with proper use of screen and view IDs.
View IDs have their own enumeration now. Their string representations
are used only during XML generation, mapping internal IDs to string
IDs.

In addition, no view is expected to send screen IDs anymore, except
for the new "error" view. Here, screen IDs are currently equivalent to
error codes.

The error view is special as it is not displayed directly on ARM side.
Instead, it is recognized as error state and may (or may not) be
displayed in any way the ARM desires.

Screen IDs are global IDs and are independent of views. Thus, it would
still be possible to send screen IDs (i.e., error IDs) with any view.
Anyway, using an explicit error view for the purpose of reporting
errors makes a lot more sense and simplifies matters on our side.

Change-Id: I04095dd63bca846e3111e381fd9679f0ce22c8d1