Replace some instances of std::bind() by lambdas.
Lambdas are a language feature whereas std::bind() is an STL feature.
Common wisdom advises to prefer lambda expressions over std::bind()
where possible since the compiler will have an easier time to perform
optimizations, and because std::bind() implementations tend to be a
big mess of template code.

Being an STL template, std::bind() might be a tad slower than lambdas,
but this strongly depends on compiler version. For sure, however, the
compiler errors are much easier to grasp with lambdas than with
std::bind().

I'll slowly replace the remaining instances of std::bind() by lambda
expressions when there is the opportunity to do this, for the reasons
stated above.

Change-Id: Id42c5811ac14de82df3c6a5d927d581dd6be0135