Manage runtime data for resuming playback.
This commit adds data to each audio source for keeping track of
position. It may seem odd to associate this data directly with audio
sources objects, but it is simple and does not try to introduce extra
abstractions or unneeded decoupling. Resume data is stored on file for
each audio source.

Change-Id: Id310bbf3c9479d32c15ea1f23d3f44ba16afc123