Reduce amount of meta data sent to SPI slave.
We send either a triple consisting of artist, track, and album, or a
triple of three pre-formatted lines (line0, line1, line2), but not
both at the same time. In case any of the artist, track, or album are
known, or in case all values are empty, the former triple is
preferred and the latter is dropped.

Since we are now preselecting/dropping data, the flexibility in what
the SPI slave could do with all those data is somewhat reduced. In
case tags from both triples are set (a case which cannot occur at the
time of this writing), the SPI slave would only see half of the truth.

Change-Id: I74464e20d74283bb717294ab8b3d1e77cbe8ce12