Suppress enforcing play mode intentions for Roon.
We are usually aggressively trying to fulfill the user's intention
(play, pause, skip, etc.). The intention is taken from the commands
sent by the remote control.

In the case of Roon, however, there is another controller for the Roon
player (the Roon app), which can also alter the state of the player
(correctly so, and by the user's intention). Thus, our own fix-up code
is disabled for Roon to allow this to happen.

Change-Id: I8ef1a8d140f6b1e96c10f61a82f757629e331358