Replace is-browse-view flag in ViewIface by set of flags.
We'll need a few more flags, so let's switch to a better and more
manageable representation right now.

As a first improvement introduced by this change, the management of
browse vs non-browse views is not based on the fact whether or not the
involved views are browse views anymore. Instead, there are two
separate feature bits now which define how the view manager should
behave, detached from any "browse view" concept and implicit,
hard-coded behaviors.

Change-Id: I58696874e341f74195dc814bdd5dec3407865578