Update view titles when switching Airable audio sources (#916).
The title in the DRC XML was not correctly updated when switching
between audio sources associated with Airable. This is because the
Airable file browser view manages all the Airable audio sources using
a shared state.

Navigational state for the individual audio sources is stored in
ViewFileBrowser::AirableView::StoredPosition objects, and the shared
state is store to and restored from these objects when switching
between the audio sources. The list title was missing from these
objects, so the view title was only never updated when simply
switching between Airable audio sources; it was only updated in case
actual D-Bus traffic had to be made and the correct title was
delivered from that path.

Change-Id: I2137e1f16753a44e76ce8983272a0634cf2909d1