Audio sources: Fix selection protocol implementation.
The D-Bus call completions for audio source selection and deselection
must be sent back to the audio source manager only after we are done
attaching to/detaching from the player (audio path manager waits for completion
and synchronizes its actions based on completion, so time of completion must
not be implemented in a sloppy way). We didn't do this, however, and notified
completion way too early. This caused us to send stop requests to the stream
player at times when we shouldn't because audio source selection already went
on and another process already took over the player.

This ultimately led to various situations where some other player
controller waited infinitely for a start notification. Audio source selection
was generally unreliable because of this bug.

Change-Id: I67fe13a0c110eb3f08f4a0e834946fa4cc0668f3