List navigation: Rewritten entirely.
There were lots and lots of problems with D-Bus communication for
long-running requests, up to the point that the software was simply
unable to run correctly on the CM3 due to entirely different timings.

This commit changes most of the direct D-Bus calls to the list brokers
by a request/notify/fetch mechanism which frees us from thinking and
working around regular D-Bus timeouts. The patch is huge and cannot be
broken done into steps in a meaningful way, so we need to use it as
is. Unfortunately.

Note that this commit should pose a warning to those who do not know
about or ignore the impact of technical dept. All of the problems in
the previous code base were introduced before the first public
release, when pressure was high to get the product out of the door and
the code was written in a hurry. It was bad code. It lived on and was
quickly patched in various ways multiple times to make it "work", sort
of, and more things were built on top. Truth is that there was no way
to fix the code because the whole approach was fundamentally broken.

Here is the cure, and it is bound to contain new bugs. At least we can
handle them now. Tons of debugging aids have been added to support us
in the future.

The D-Bus list class saw various improvements, simplifications, and
fixes as well to integrate the new calling convention.

There were (and deep down, still are) various problems concerning
memory management for asynchronous GIO calls. It is not really
possible to use smart pointers inside a static callback function
called from GLib, at least not in a sensible way.

The GIO callback passes a raw pointer to some user data, which may or
may not have to be deallocated when called. If that pointer to user
data is supposed to be a smart pointer, then it still cannot *be* one
because the function signature simply expects a raw pointer. Of course
we still want to use smart pointers, so we need to keep them in a
different place, outside of GLib. We did this, but it turned out to be
complicated, and several different UAF bugs showed up in different
contexts.

Now we simply keep shared pointers to all the objects ever passed to
an async GIO function in a set (they are registered, roughly
equivalent to being allocated). Inside the static callback, the object
gets unregistered (which roughly resembles a free() call).

The fundamental problem, plainly spoken, is that C sucks for
application programming. And GLib, being a C library and a royal
clusterfuck in it own right, sucks even more so.

Change-Id: Iba202292012a6920c8e7b26e34ed5713aa848979