player: Remove app stream bookkeeping.
It is not necessary anymore to do this. We receive all the necessary
information from dcpd (stream info) and streamplayer (now playing,
stopped, dropped, etc.).

The information are reliable since dcpd v5.4, and they were reliable
in ~99.9% of all cases in earlier versions. Combined with information
about dropped streams added in streamplayer v5.0, we do not need to
keep track of app streams anymore.

Change-Id: I0f9302ccf79f65a0cd703fd7ad144dc53cab7918