meta data: Remove MetaData::Reformatters struct because YAGNI.
The original reason to touch this code in the first place was to avoid
a C++ compiler warning about designated initializers, which are only
supported in C99 and later, but not in C++.

The MetaData::Reformatters object used to be a POD struct with a
function inside, initialized by a designated initializers. This
corresponds to a C++ object with a single member vtable, so the
straightforward fix would have been to implement it this way. Since
the function in question is just a plain function, however, it
wouldn't have made much sense to wrap it into a class, to make an
instance of that class, and to pass the object around (like we did it
with the struct instance). The original intent, of course, was the
capability to exchange the output format based on context by injecting
different Reformatter objects, thus flexibility, but that flexibility
was never needed.

This patch replaces the MetaData::Reformatters struct by a namespace
of the same name and places the only reformatting function into that
namespace.

Change-Id: I36652844f7bad8665fe364e58a016bf9b75d6e11