file view: Avoid gcc warning about case fall-through.
We use a C++17 feature to suppress the warning because reasonably
recent versions of gcc accept it even in C++14 mode. The manpage
states that the GNU extension [[gnu:fallthrough]] could be used for
C++14 and below, but it won't work with gcc 5.2 (our current cross
compiler).

So there's an #if now... The preprocessor check should be removed when
moving to a more recent version of Yocto Project, thus gcc.

Change-Id: I98a131e949a74ca21f88a5c32c67c873fa9fa89a