rnfcall: Suppress bug messages from dtor.
The bug messages could also be suppressed by calling
DBusRNF::CallBase::abort_request() from the dtor only when abortion is
allowed, but it would unnecessarily complicate the code. Calling the
abort_request() function is, in fact, always safe, but bug messages
are emitted in case regular client code screws up. The dtor is not
regular client code, so it can live on relaxed standards.

Change-Id: I921ffe09ede61ac5104d219a83ae3059313fc273