filebrowser: Move debug output to View::write_xml().
The synchronous debug output in View::serialize() has interfered
several times with regular operation, and it has happened again.
Occasional deadlocks and interrupted operations can be accounted to
that code, so it is only executed in foreground mode now (which is
used for explicit debugging).

A new debug output has been added to View::write_xml() along with
regular, asynchronous XML serialization because such an output proved
to be incredibly useful in the past.

Change-Id: Ia262b47afe90a1f84f0dfe94baeb0c9c748411f0