dbuslist: Synchronize start and end of multiple GetRange ops (#1267).
There was a problem with the arrangement of operations for filling in
the D-Bus list's internal cache window. Our old D-Bus list code relies
on certain operations to work and complete in a specific order, and
the fundamental assumption that only a single D-Bus operation could
take place at any given point in time. This assumption is no longer
true, and multiple operations may run concurrently now. This breaks
the ordering of operations, especially their completion, because the
final handling of GetRange operations was deferred to the main context
(this was done to circumvent locking problems and deep recursions).

Because of different ordering, crashes were observed when navigating a
slow list very quickly. In this situation, GetRange operations are
frequently canceled and replaced by new operations. The cancellations
were handled late and in bulk after the successful GetRange operation
has completed, but the cancellations also messed up the list's
internal cache window. It was pretty difficult to find the root
problem for this.

A good solution would have been to ditch the whole caching strategy
and replace it with code which can cope with multiple, concurrent
D-Bus operations. Such code would be easier to maintain and faster in
practice, it would, however, also mean a major rewrite of the D-Bus
list code.

Thus, this commit does the second best thing, which is serialization
of operations to match the ordering of the old code. Operations are
completed before another operation can continue. This is not optimal,
but the implementation was fairly easy, in retrospect.

This commit also changes a few other unrelated things which I am not
going to split into separate commits. First, the
List::CacheSegment::Intersection enumeration as been replaced by the
List::SegmentIntersection class enumeration. Second, the
List::CacheSegment class has been renamed to List::Segment and its
public data members are private now. A few function members have been
added to simplify client code. Third, many function signatures have
been changed so that line/count parameter pairs are List::Segment
references now to make the code more comprehensible.

Change-Id: I8aca9e931b1a7c0df4d90c6493889983631c6360